Incorporated code review suggestions
